Move filebased routes out on their own
